+++ subs2cia/tests/ffmpeg_tools_audiotrim_test.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from subs2cia.ffmpeg_tools import *
 import base64
 
+
 def get_args():
     parser = argparse.ArgumentParser(description=f'audio trimming manual testing')
 
@@ -33,8 +34,10 @@ if __name__ == '__main__':
     # ffmpeg_trim_audio_clip_encode(Path(args.videofile), stream_index=2, timestamp_start=40000, timestamp_end=50000, quality=None,
     #                        to_mono=True, outpath=Path("output.mp3"))
 
-    stdout = ffmpeg_trim_audio_clip_atrim_encode(Path(args.videofile), stream_index=0, timestamp_start=40000, timestamp_end=50000, quality=None,
-                           to_mono=True, normalize_audio=True, outpath=Path("pipe:"), format="mp3", capture_stdout=False)
+    stdout = ffmpeg_trim_audio_clip_atrim_encode(Path(args.videofile), stream_index=0, timestamp_start=40000,
+                                                 timestamp_end=50000, quality=None,
+                           to_mono=True, normalize_audio=True, outpath=Path("pipe:"),
+                                                 format="mp3", capture_stdout=True, silent=False)
     encoded = base64.b64encode(stdout)
     print(encoded)
     decoded = base64.b64decode(encoded)